Refactor
Add subfolders to comonponet dir to group components based on their pages
Break pages into at least one componet
Add semantic HTML
Add margin bottom to slug page
