+++ components/layout/Footer.js
@@ -1,0 +1,7 @@
+export default function Footer() {
+  return (
+    <footer className="absolute bottom-0 flex h-16 w-full items-center justify-center  bg-black text-white">
+      <p>Sandals Church &copy; {new Date().getFullYear()}</p>
+    </footer>
+  )
+}
