+++ functions/functions.js
@@ -1,3 +1,4 @@
+//Safari and IE cannot parse date strings. Must convert to numbers pass in as Year, month, day. Example: new Date(2011, 01, 07)
 export function parseDate(date) {
   const [year, month, day] = date.split('-')
