+++ pages/_app.js
@@ -1,5 +1,5 @@
 import '../styles/globals.css'
-import Layout from '../components/Layout'
+import Layout from '../components/layout/Layout'
 
 function MyApp({ Component, pageProps }) {
   return (
