+++ tsconfig.json
@@ -19,7 +19,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    "components/Header.js",
+    "components/index/Header.js",
     "components/Navbar.tsx",
     "pages/index.js",
     "pages/_app.js"
